Change costs type from int to double in AssignmentService.Solve.

+++ PCSetupHub.Core/Interfaces/IAssignmentService.cs
@@ -8,6 +8,6 @@
 		/// <param name="costs">Matrix of assignment costs.</param>
 		/// <param name="findMax">True to maximize total cost, false to minimize.</param>
 		/// <returns>Array mapping each agent to its assigned task.</returns>
-		int[] Solve(int[,] costs, bool findMax);
+		int[] Solve(double[,] costs, bool findMax);
 	}
 }
